Create backups of KJots books on close (max 10 backup versions). Remove backups of deleted books 7 days later. That gives you a week to realise you shouldn't have deleted the book and recover it.
http://userbase.kde.org/Tutorials/Recover_Deleted_KJots_Books

BUG: 175982


svn path=/trunk/KDE/kdepim/; revision=890816

+++ kjots/kjotsentry.cpp
@@ -373,6 +373,15 @@ void KJotsBook::saveBook(void)
     }
 }
 
+void KJotsBook::saveAndBackupBook()
+{
+    if (!m_fileName.isEmpty())
+    {
+      KSaveFile::numberedBackupFile(m_fileName);
+    }
+    saveBook();
+}
+
 /*!
     \brief Deletes a book by removing the data file.
     This does not affect the list display, and can be called for reasons other than
@@ -381,7 +390,10 @@ void KJotsBook::saveBook(void)
 */
 void KJotsBook::deleteBook ( void )
 {
-    QFile::remove(m_fileName);
+    if (!m_fileName.isEmpty())
+    {
+        QFile::remove(m_fileName);
+    }
     m_fileName.clear();
 }
